% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app.R
\name{app_SCRAD}
\alias{app_SCRAD}
\title{SCRAD v0.1}
\usage{
app_SCRAD()
}
\value{
A shiny app to visualization
}
\description{
A shiny app to visualizate the risk of SNPs associated with five types of dementia
}
\examples{
app_SCRAD()

}
